requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
cx-Freeze==6.6
cx-Logging==3.0
DAWG-Python==0.7.2
discord.py==1.7.3
docopt==0.6.2
emoji==1.2.0
idna==2.10
importlib-metadata==4.5.0
multidict==5.1.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
zipp==3.4.1
